Build: Allow building without cython installed (based on pre generated mpfr.c)
In case cython is not installed on the system, fallback on building the
wheel with the mpfr.c.

This file was generated and added to the git (instead of being re
generated thanks to cython).

This commits helps to express dependencies of a package, on bigfloat, from github repo.
Without the guarantie that Cython is installed on the systeme.